Dao+DaoImpl V2

+++ src/main/java/ru/tsystems/mis/spring/dao/implementations/OptionDAOImpl.java
@@ -2,7 +2,10 @@ package ru.tsystems.mis.spring.dao.implementations;
 
 import org.hibernate.Session;
 import org.hibernate.SessionFactory;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 import ru.tsystems.mis.spring.dao.interfaces.OptionDAO;
 import ru.tsystems.mis.spring.model.Option;
@@ -10,56 +13,65 @@ import ru.tsystems.mis.spring.model.Option;
 import java.io.Serializable;
 import java.util.List;
 
+@Repository
 public class OptionDAOImpl implements OptionDAO {
+
+    private static final Logger logger = LoggerFactory.getLogger(OptionDAOImpl.class);
 
     @Autowired
     SessionFactory factory;
 
     @Transactional
     @Override
-    public int add(Option option) {
+    public void addOption(Option option) {
         Session session = factory.openSession();
-        Serializable id = session.save(option);
+        session.save(option);
+        logger.info("Option successfully saved. Option details: " + option);
         session.close();
-        return Integer.valueOf(id.toString());
     }
 
     @Transactional
     @Override
-    public int update(Option option) {
+    public void updateOption(Option option) {
         Session session = factory.openSession();
         session.update(option);
-        Serializable id = session.getIdentifier(session);
+        logger.info("Option successfully update. Option details: " + option);
         session.close();
-        return Integer.valueOf(id.toString());
     }
 
 
     @Transactional
     @Override
-    public int delete(Option option) {
+    public void deleteOption(Long id) {
         Session session = factory.openSession();
-        session.delete(option);
-        Serializable id = session.getIdentifier(session);
+        Option option = (Option) session.load(Option.class, new Long(id));
+        if(option != null){
+            session.delete(option);
+            logger.info("Option successfully delete. Option details: " + option);
+        }
         session.close();
-        return Integer.valueOf(id.toString());
+
     }
 
     @Transactional
     @Override
-    public Option get(Long id) {
+    public Option getOptionById(Long id) {
         Session session = factory.openSession();
-        Option option = session.load(Option.class, id);
+        Option option = (Option) session.load(Option.class, new Long(id));
+        logger.info("Option successfully loaded. Option details: " + option);
         session.close();
         return option;
     }
 
     @Transactional
     @Override
-    public List<Option> list() {
+    @SuppressWarnings("unchecked")
+    public List<Option> listOptions() {
         Session session = factory.openSession();
-        @SuppressWarnings("unchecked")
         List<Option> list = session.createQuery("from Option").list();
+        for (Option option : list) {
+            logger.info("Option list: " + option);
+        }
         session.close();
         return list;
     }
